+++ backend/hqlib/metric_source/abstract/coverage_report.py
@@ -14,11 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-
-import datetime
 import functools
-import bs4
-from typing import Callable, List
+from typing import Callable
 
 from ..url_opener import UrlOpener
 from ... import domain
@@ -33,57 +30,47 @@ class CoverageReport(domain.MetricSource):
         self.__url_open = url_open or UrlOpener(**kwargs).url_open
         super().__init__()
 
-    def statement_coverage(self, coverage_url: str) -> int:
-        """ Return the ART statement coverage for a specific product. """
-        try:
-            soup = self.__get_soup(coverage_url)
-        except UrlOpener.url_open_exceptions:
-            coverage = -1
-        else:
-            coverage = self._parse_statement_coverage_percentage(soup)
-        return coverage
-
-    def _parse_statement_coverage_percentage(self, soup) -> int:
-        """ Parse the coverage percentage from the soup. """
+    def statement_coverage(self, metric_source_id: str) -> float:
+        """ Return the statement coverage for a specific product. """
         raise NotImplementedError
 
-    def branch_coverage(self, coverage_url: str) -> int:
-        """ Return the ART branch coverage for a specific product. """
-        try:
-            soup = self.__get_soup(coverage_url)
-        except UrlOpener.url_open_exceptions:
-            coverage = -1
-        else:
-            coverage = self._parse_branch_coverage_percentage(soup)
-        return coverage
-
-    def _parse_branch_coverage_percentage(self, soup) -> int:
-        """ Parse the coverage percentage from the soup. """
+    def branch_coverage(self, metric_source_id: str) -> float:
+        """ Return the branch coverage for a specific product. """
         raise NotImplementedError
 
     @functools.lru_cache(maxsize=1024)
-    def datetime(self, *coverage_urls: str) -> DateTime:
-        """ Return the date when the ART coverage for a specific product was last successfully measured. """
-        coverage_date_urls = self._get_coverage_date_urls(coverage_urls[0])
-        for index, coverage_date_url in enumerate(coverage_date_urls):
-            is_last_url = index + 1 == len(coverage_date_urls)
-            try:
-                soup = self.__get_soup(coverage_date_url, log_error=is_last_url)
-            except UrlOpener.url_open_exceptions:
-                continue
-            return self._parse_coverage_date(soup)
-        return datetime.datetime.min
-
-    def _parse_coverage_date(self, soup) -> DateTime:
-        """ Parse the coverage date from the soup. """
+    def datetime(self, *metric_source_ids: str) -> DateTime:
+        """ Return the date when the coverage for a specific product was last successfully measured. """
         raise NotImplementedError
 
-    @staticmethod
-    def _get_coverage_date_urls(coverage_url: str) -> List[str]:
-        """ Return the url for the date when the coverage of the product was last measured. """
-        return [coverage_url]
+
+class UnittestCoverageReport(CoverageReport):
+    """ Abstract class representing a unit test coverage report. """
+    def statement_coverage(self, metric_source_id: str) -> float:
+        """ Return the statement coverage for a specific product. """
+        raise NotImplementedError
+
+    def branch_coverage(self, metric_source_id: str) -> float:
+        """ Return the branch coverage for a specific product. """
+        raise NotImplementedError
 
     @functools.lru_cache(maxsize=1024)
-    def __get_soup(self, url: str, log_error=True):
-        """ Get a beautiful soup of the HTML at the url. """
-        return bs4.BeautifulSoup(self.__url_open(url, log_error=log_error), "lxml")
+    def datetime(self, *metric_source_ids: str) -> DateTime:
+        """ Return the date when the coverage for a specific product was last successfully measured. """
+        raise NotImplementedError
+
+
+class ARTCoverageReport(CoverageReport):
+    """ Abstract class representing a unit test coverage report. """
+    def statement_coverage(self, metric_source_id: str) -> float:
+        """ Return the statement coverage for a specific product. """
+        raise NotImplementedError
+
+    def branch_coverage(self, metric_source_id: str) -> float:
+        """ Return the branch coverage for a specific product. """
+        raise NotImplementedError
+
+    @functools.lru_cache(maxsize=1024)
+    def datetime(self, *metric_source_ids: str) -> DateTime:
+        """ Return the date when the coverage for a specific product was last successfully measured. """
+        raise NotImplementedError