+++ backend/hqlib/metric_source/test_report/sonar_test_report.py
@@ -21,8 +21,8 @@ from hqlib.typing import DateTime
 
 
 class SonarTestReport(metric_source.TestReport):
-    """ Metric source for retrieving (unit) test data from Sonar. """
-    metric_source_name = "Sonar test report"
+    """ Metric source for retrieving (unit) test data from SonarQube. """
+    metric_source_name = "SonarQube test report"
 
     def __init__(self, sonar_url: str, *args, **kwargs) -> None:
         sonar_class = kwargs.pop('sonar_class', metric_source.Sonar)