+++ backend/tests/unittests/metric_source/coverage_report/ncover_tests.py
@@ -55,7 +55,7 @@ class NCoverTest(unittest.TestCase):
 
     def setUp(self):
         self.__opener = FakeUrlOpener()
-        NCover._CoverageReport__get_soup.cache_clear()
+        NCover._HTMLCoverageReport__get_soup.cache_clear()
         NCover.datetime.cache_clear()
         self.__ncover = NCover(url_open=self.__opener.url_open)